pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "box_orientation"
version = "0.1.0"
description = "Multi-camera application for determining the orientation of boxes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "framegrab>=0.10.0",
    "groundlight>=0.22.3",
    "imgcat>=0.5.0",
    "matplotlib>=3.10.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/box_orientation"]
